fix interrupt handling

+++ support.c
@@ -166,8 +166,8 @@ static void set_idt_entry( int entry, void ( *handler )( void ) ){
 	g->offset_15_0 = (unsigned short)handler & 0xffff;
 	g->segment_selector = 0x0010;
 	g->flags = IDT_PRESENT | IDT_DPL_0 | IDT_INT32_GATE;
-	g->offset_31_16 = (unsigned short)handler >> 16 & 0xffff;
-	g->offset_63_32 = (unsigned long long)handler >> 32 & 0xffffffff;
+	g->offset_31_16 = ((unsigned int)handler >> 16) & 0xffff;
+	g->offset_63_32 = ((unsigned long long)handler >> 32) & 0xffffffff;
 	g->must_be_zero = 0;
 }
